requirements.txt
pyserial
smbus
colorama